+++ src/angularjs/src/app/home/neighborhood-map.directive.js
@@ -9,7 +9,7 @@
         ctl.$onInit = function () {
             ctl.mapOptions = { scrollWheelZoom: false };
             ctl.boundsConus = MapConfig.conusBounds;
-            ctl.baselayer = L.tileLayer(
+            ctl.baseLayer = L.tileLayer(
                 MapConfig.baseLayers.Positron.url, {
                     attribution: MapConfig.baseLayers.Positron.attribution,
                     maxZoom: MapConfig.conusMaxZoom