+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 #### Added
 - A delete button for neighborhoods, with a confirmation modal
+- Clean up S3 assets when deleting neighborhoods and analysis jobs
 
 ## [0.11.0] - 2019-10-31